Renamed some things and removed uneeded directories.
