+++ .gitignore
@@ -33,3 +33,6 @@
 
 # DS_Store
 .DS_Store
+
+build/
+cmake-build-debug/
