+++ include/sproj_2024/gui/FreeDeckGUI.h
@@ -3,9 +3,9 @@
 //
 #pragma once
 #include <juce_gui_extra/juce_gui_extra.h>
-#include <DeckGUI.h>
-#include <DummyClip.h>
-#include <helpers.h>
+#include "DeckGUI.h"
+#include "DummyClip.h"
+#include <sproj_2024/helpers/helpers.h>
 
 class FreeDeckGUI final : public juce::Component,
                           public DeckGUI,
