+++ include/sproj_2024/gui/MenuComponent.h
@@ -5,8 +5,8 @@
 #pragma once
 
 #include <juce_gui_basics/juce_gui_basics.h>
-#include <helpers.h>
-#include <SPCommandManager.h>
+#include <sproj_2024/helpers/helpers.h>
+#include "SPCommandManager.h"
 
 class MenuComponent final : public juce::Component,
                             public juce::MenuBarModel,
