+++ include/sproj_2024/gui/RulerDeckGUI.h
@@ -3,7 +3,7 @@
 //
 #pragma once
 #include <juce_gui_extra/juce_gui_extra.h>
-#include <DeckGUI.h>
+#include "DeckGUI.h"
 
 class RulerDeckGUI final : public juce::Component,
                           public DeckGUI
