+++ include/sproj_2024/gui/SPCommandManager.h
