+++ src/Main.cpp
@@ -1,8 +1,12 @@
-#include <MainAudio.h>
-#include "MainWindow.h"
-#include "helpers.h"
-#include <SPCommandManager.h>
-#include "focusrite/e2e/TestCentre.h"
+/**
+ * Entry point
+ */
+
+#include <sproj_2024/audio/MainAudio.h>
+#include <sproj_2024/gui/MainWindow.h>
+#include <sproj_2024/helpers/helpers.h>
+#include <sproj_2024/gui/SPCommandManager.h>
+#include <focusrite/e2e/TestCentre.h>
 
 /* TODO:
  *
