+++ test/unit_tests/AudioClockTest.cpp
@@ -3,11 +3,11 @@
 //
 
 #include <gtest/gtest.h>
-#include "sproj_2024_lib/audio/Track.h"
-#include <sproj_2024_lib/gui/FreeDeckGUI.h>
-#include <sproj_2024_lib/gui/SPCommandManager.h>
-#include <sproj_2024_lib/audio/AudioClock.h>
-#include <sproj_2024_lib/audio/MainAudio.h>
+#include <sproj_2024/audio/Track.h>
+#include <sproj_2024/gui/FreeDeckGUI.h>
+#include <sproj_2024/gui/SPCommandManager.h>
+#include <sproj_2024/audio/AudioClock.h>
+#include <sproj_2024/audio/MainAudio.h>
 
 TEST(AudioClockTest, Clock_PauseOrResumeProcessing_1_Test)
 {
